All activities are functional and streamlined
All the quest activities (multiple choice, dragging, and fill in the
blank) are working, as are all of the database functions. The code has
been streamlined and commented.

The only thing not working is the masking function in bag.lua. That is
a Corona SDK bug that needs to get fixed